+++ tests/testthat/test_lambert.R
@@ -20,6 +20,23 @@ test_that('lambert Transforms new data consistently', {
   expect_equal(nd, nd2)
 })
 
+lambert_obj <- lambert(train, standardize = FALSE)
+
+test_that('lambert without standardization Transforms original data consistently' , {
+  expect_equal(lambert_obj$x.t, predict.lambert(lambert_obj))
+  expect_equal(lambert_obj$x, predict.lambert(lambert_obj, inverse = T))
+})
+
+test_that('lambert without standardization Transforms new data consistently', {
+  nd <- seq(-1, 4, length = 100)
+  pred <- predict(lambert_obj, newdata = nd)
+  expect_true(!any(is.na(pred)))
+  
+  nd2 <- predict(lambert_obj, newdata = pred, inverse = TRUE)
+  expect_equal(nd, nd2)
+})
+
+
 # For type = 'hh'
 
 lambert_obj <- lambert(train, type = 'hh')